favorite

+++ src/main/java/com/jjim/drawer/application/DrawerService.java
@@ -4,8 +4,9 @@ import com.jjim.common.domain.PagedResponse;
 import com.jjim.common.exception.drawer.DrawerException;
 import com.jjim.common.exception.drawer.DrawerExceptionCode;
 import com.jjim.drawer.domain.entity.Drawer;
-import com.jjim.drawer.infra.DrawerRepository;
+import com.jjim.drawer.infra.repository.DrawerRepository;
 import com.jjim.drawer.presentation.response.DrawerResponse;
+import com.jjim.favorite.infra.repository.FavoriteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -20,6 +21,7 @@ import java.util.UUID;
 public class DrawerService {
 
     private final DrawerRepository drawerRepository;
+    private final FavoriteRepository favoriteRepository;
 
     @Transactional(readOnly = true)
     public PagedResponse<DrawerResponse> getDrawers(final UUID userId, final Pageable pageable) {
@@ -52,9 +54,14 @@ public class DrawerService {
 
     @Transactional
     public void deleteDrawer(final UUID userId, final Long drawerId) {
-        final Drawer drawer = drawerRepository.findByUserIdAndId(userId, drawerId)
+        final Drawer drawer = drawerRepository.findById(drawerId)
             .orElseThrow(() -> new DrawerException(DrawerExceptionCode.DRAWER_NOT_FOUNT));
 
+        if (!drawer.getUserId().equals(userId)) {
+            throw new DrawerException(DrawerExceptionCode.DRAWER_NOT_MATCH_USER);
+        }
+
+        favoriteRepository.deleteFavoritesByDrawerId(drawerId);
         drawerRepository.delete(drawer);
 
     }
